cambie el nombre de las clase carnetdetalle por detallepago y conecte los layout
