--- app/src/main/java/com/deportes/clubdeportivo/CarnetDetalleActivity.kt
@@ -1,22 +1,0 @@
-package com.deportes.clubdeportivo
-
-import android.os.Bundle
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-
-class CarnetDetalleActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_carnet_detalle)
-
-        // Importación y lógica de la barra superior
-        val btnAtras: ImageView = findViewById(R.id.buttonBack)
-        val textViewTitulo: TextView = findViewById(R.id.textViewTitle)
-        textViewTitulo.text = "Carnet"
-
-        btnAtras.setOnClickListener {
-            finish()
-        }
-    }
-}
